add delete confirmation page

+++ products/views.py
@@ -143,6 +143,10 @@ def delete_product(request, product_id):
         return redirect(reverse('home'))
 
     product = get_object_or_404(Products, pk=product_id)
-    product.delete()
-    messages.success(request, f'{product.name} deleted!')
-    return redirect(reverse('products'))
+
+    if request.method == 'POST':
+        product.delete()
+        messages.success(request, f'{product.name} deleted!')
+        return redirect(reverse('products'))
+
+    return render(request, 'products/confirm_delete.html')
